Create client class diagram; refactoring
