+++ doc/client-class.drawio
@@ -1,0 +1,1 @@
+<mxfile host="app.diagrams.net" modified="2020-04-02T11:46:25.883Z" agent="5.0 (Windows NT 6.1; Win64; x64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/80.0.3987.162 Safari/537.36" etag="JT4eOj8xBNSNH3EZFXOQ" version="12.9.6" type="device"><diagram id="z5wTcfgIbSRSx7cCiWtq" name="Page-1">5Ztdc5s4FIZ/jS/jAfFh+zJx0nZ3prudujtNLgnItiYyYoVc2/vrVwTJIAkbQjB40puMJeAArx4dnXNERs58s/9Mg2T9lUQQj4AV7UfO/QiAqT/lf7OOQ97hTkHesaIoyrvsomOB/oOi0xK9WxTBVDmREYIZStTOkMQxDJnSF1BKduppS4LVuybBChodizDAZu9PFLG1eC0wKfq/QLRayzvb/iw/sgnkyeJN0nUQkV2py3kYOXNKCMt/bfZziDPtpC75dZ9OHD0+GIUxa3LBd7j7N8Kf/8aPP7cvNzfuMv6BbsRg/ArwVrzwHCNucMG2z+Kx2UFqQck2jmBmzho5d7s1YnCRBGF2dMcHn/et2Qbzls1/CsOQMrg/+cT2UQfODyQbyOiBnyIu8IVyAh3g5c1dMQ62FHddGgN5WSCGfnU0XKjDfwiB3iCWc0KsPzYJHlwsMJspatmeNbBcriHXV5imfMLco5AhEgfiHkOq5lkqY44/tGqeCRmhMAdtcLl0yMBsaLl8Q64vjF2pWi4YWq2JodZdkKLwNmSEDi6XfXVzcXrC4X8nZDO4XLrrAoPTNTPkyoT6MyXxYpskhF7fjPTcoTWTcW5JtNEcjG7vUMwgXWZa8OYdf3kf82e5e+bz1F+xVzlEj+yQXi/v5k/zXJyq6d6tqo7lqhO3QlS7QlTvYqLapqi6BDCObrP8gLciFGxIHP1Yo5gLwg98QljKw1siKQGuiiMXhx4ey42njOOxJ5v3e8F13jrI1h6xx9Lv0lW8VVyUNeQ1+dPDyMhVtCHib0i2NIQNgGMBXUF25kRQPeblQZWzgkIcMPRLfbiqURXmvhGUgXqclZonO8IiTeQvJa4qpziaIcevMZS/tGHoFbLjO76DOzOdOsfdMybhCx/dKEjXr07QVvCzFPy4bRU/idERnacyOZUYHZEtUfpUJvgEsh3i5zSkz7Z6w8+eadS4bfHzagxdGj8zQb2A22vBXYf8eA35cfrDZ6rFYa3xcWsMXRofM2G/DnyuwG35DbHz+sNOo8XRQ9TG2J1Ku/rCzqx4POwZJyhtsnaWFkjfWD2ruHtD1HWKO6s/7pq6u0l/3Nmal2rLnQEwaMYdJyA4lE5LshPSMw/sV9+nwDi32C3UZl3q+qB2XaC407FVm7tkrW+QIq4PpMO52f5w1xZ3ty3twK0x1BXtnqPex+uDdrOueJskGIVBVud/rS4uDimDZtGs65KY8u6VWyCyT6nueKexeFchQrq063MDHU7axildj4uUXht12k5bfQ9SN3Th4AiYlaxLAqQE5XZfADX1+rPe+PGAFjO0zen8aY2hS/NjVqT+SeE5dkgCY7PwpJSoKsix20Ug1x5Wu70R52prl9O2BurbNYY6CjS8IcJqYFa4rhbnK6xO9Iez52jh7qQtzrMaQxfCuZe4GZgVN7mJZy2DMIhMtNN1kGQ/Y8Jg/ZZpKpjPinBZZM0jcnyL0YpPiXtGspMD0cJwme0UptwUilc/smP3N343Ybmru4qKsNyqCMv1tbK7sNzMzs/5kJjE8IzTaD6R6/fc6mdoVf7idbQGyQqwnAOzdnNWn/w9FxSB+dnGudF95+6vNZ68KYoZOFmbNiSxv0L20aPLrL31UuHUGOpoqXCm2kdwvSwV5sc1csgLqrmLZtoCwCh5gXOCCS0c2ZIDrnXJZSCE8Wthz1gtNiiK8KkVR/2Mp4MVQ9+SrVgx3Ao/qFeQu/vy16zjWOPx+K+POwLAr/2mp98RMAshH5h/R6s/ANv89r1f+c06wu8k/2Ro+s2894P7Hz1nGN4BmbnaR54BM9X9Dz8DzC2mjyy/Xo2whpbfzKl+I/mdio3EjuTnzeJfAPNkofg/Sufhfw==</diagram></mxfile>
