+++ src/main/scala/client/ClientStub.scala
@@ -8,11 +8,6 @@ object ClientStub extends App with ApplicationActorSystem{
   private val serverAddress = "localhost"
   private val serverPort = 8080
 
- /* private val gameServer = GameServer(serverAddress, serverPort)
-  gameServer.defineRoom("test_room", RoomStrategy.empty)
-  Await.ready(gameServer.start(), 5 seconds)
-  println(s"Server started at $serverAddress:$serverPort")*/
-
   val client = Client(serverAddress, serverPort)
   client createPublicRoom("test_room", "") onComplete {
     case Success(room) => println("Room id ->" + room.roomId)
