+++ src/main/scala/client/CoreClient.scala
@@ -23,7 +23,7 @@ class CoreClientImpl(private val serverUri: String) extends CoreClient with Room
   private val httpClient = context.system actorOf HttpClient(serverUri, self)
   private var joinedRooms: Set[ClientRoom] = Set()
 
-  import MessageDictionary._
+  import client.utils.MessageDictionary._
 
   val onReceive: Receive = {
     case CreatePublicRoom(roomType, _) =>
