+++ src/main/scala/common/RoomJsonSupport.scala
@@ -2,7 +2,7 @@ package common
 
 import akka.http.scaladsl.marshallers.sprayjson.SprayJsonSupport
 import client.room.ClientRoom.ClientRoom
-import common.CommonRoom.{Room, RoomId}
+import common.SharedRoom.{Room, RoomId}
 import spray.json.{DefaultJsonProtocol, JsArray, JsString, JsValue, RootJsonFormat, deserializationError}
 import common.BasicRoomPropertyValueConversions._
 
@@ -39,7 +39,7 @@ trait RoomJsonSupport extends SprayJsonSupport with DefaultJsonProtocol {
   }
 
   //TODO: only works with RoomProperty[Any]
-  implicit val simpleRoomPropertyJsonFormat: RootJsonFormat[RoomProperty] = new RootJsonFormat[RoomProperty] {
+  implicit val simpleRoomPropertyJsonFormat: RootJsonFormat[RoomProerty] = new RootJsonFormat[RoomProperty] {
     def write(a: RoomProperty): JsValue = JsArray(JsString(a.name), JsString(a.value.toString))
 
     import spray.json.deserializationError
