+++ src/test/scala/client/CoreClientSpec.scala
@@ -8,7 +8,7 @@ import akka.testkit.{ImplicitSender, TestKit}
 import com.typesafe.config.ConfigFactory
 import org.scalatest.{BeforeAndAfter, BeforeAndAfterAll}
 import org.scalatest.wordspec.AnyWordSpecLike
-import client.MessageDictionary._
+import client.utils.MessageDictionary._
 import client.room.ClientRoom.ClientRoom
 import common.Routes
 import org.scalatest.matchers.should.Matchers
